Renamed Capablanca80 as Capablanca
