+++ tests/unit/Media/BoardToJpgTest.php
@@ -5,7 +5,7 @@ namespace Chess\Tests\Unit\Media;
 use Chess\Media\BoardToJpg;
 use Chess\Player\PgnPlayer;
 use Chess\Tests\AbstractUnitTestCase;
-use Chess\Variant\Capablanca80\Board as Capablanca80Board;
+use Chess\Variant\Capablanca\Board as CapablancaBoard;
 use Chess\Variant\Classical\FEN\StrToBoard as ClassicalFenStrToBoard;
 use Chess\Variant\Classical\Board as ClassicalBoard;
 
@@ -157,7 +157,7 @@ class BoardToJpgTest extends AbstractUnitTestCase
      */
     public function output_start_capablanca80()
     {
-        $board = new Capablanca80Board();
+        $board = new CapablancaBoard();
 
         $filename = (new BoardToJpg($board))->output(self::OUTPUT_FOLDER);
 
@@ -172,7 +172,7 @@ class BoardToJpgTest extends AbstractUnitTestCase
      */
     public function output_capablanca80_Nj3_e5___Ci6_O_O()
     {
-        $board = new Capablanca80Board();
+        $board = new CapablancaBoard();
 
         $board->play('w', 'Nj3');
         $board->play('b', 'e5');
