+++ tests/unit/Variant/Capablanca/FEN/BoardToStrTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Chess\Tests\Unit\Variant\Capablanca80\FEN;
+namespace Chess\Tests\Unit\Variant\Capablanca\FEN;
 
 use Chess\Variant\Classical\FEN\BoardToStr;
 use Chess\Tests\AbstractUnitTestCase;
-use Chess\Variant\Capablanca80\Board;
+use Chess\Variant\Capablanca\Board;
 
 class BoardToStrTest extends AbstractUnitTestCase
 {
