+++ tests/unit/Variant/Capablanca/PGN/AN/SquareTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Chess\Tests\Unit\Variant\Capablanca80\PGN\AN;
+namespace Chess\Tests\Unit\Variant\Capablanca\PGN\AN;
 
 use Chess\Exception\UnknownNotationException;
 use Chess\Tests\AbstractUnitTestCase;
-use Chess\Variant\Capablanca80\PGN\AN\Square;
+use Chess\Variant\Capablanca\PGN\AN\Square;
 
 class SquareTest extends AbstractUnitTestCase
 {
